Refactor sidebar components
