+++ src/lib/components/sidebar/Projects.svelte
