+++ src/routes/(app)/upload/+page.svelte
@@ -1,14 +1,11 @@
 <script lang="ts">
-  import { page } from "$app/stores";
   import { _ } from "svelte-i18n";
 
-  import SidebarLayout from "@/lib/components/layouts/SidebarLayout.svelte";
-  import SignedIn from "@/lib/components/common/SignedIn.svelte";
-
-  import Documents from "../documents/sidebar/Documents.svelte";
-  import Projects from "../documents/sidebar/Projects.svelte";
-
+  import SignedIn from "$lib/components/common/SignedIn.svelte";
   import DocumentUpload from "$lib/components/forms/DocumentUpload.svelte";
+  import SidebarLayout from "$lib/components/layouts/SidebarLayout.svelte";
+  import Documents from "$lib/components/sidebar/Documents.svelte";
+  import Projects from "$lib/components/sidebar/Projects.svelte";
 
   export let data;
